+++ tests/test_session.py
@@ -6,12 +6,12 @@ from unittest.mock import patch
 import responses
 
 from amazonorders.exception import AmazonOrdersAuthError
-from amazonorders.session import AmazonSession, BASE_URL
+from amazonorders.session import AmazonSession, BASE_URL, SIGN_IN_URL
 from tests.unittestcase import UnitTestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2024, Alex Laird"
-__version__ = "1.0.5"
+__version__ = "1.0.6"
 
 
 class TestSession(UnitTestCase):
@@ -46,7 +46,7 @@ class TestSession(UnitTestCase):
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-invalid-email.html"), "r", encoding="utf-8") as f:
             resp2 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
@@ -73,7 +73,7 @@ class TestSession(UnitTestCase):
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-invalid-password.html"), "r", encoding="utf-8") as f:
             resp2 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
@@ -101,14 +101,14 @@ class TestSession(UnitTestCase):
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-mfa.html"), "r", encoding="utf-8") as f:
             resp2 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
         with open(os.path.join(self.RESOURCES_DIR, "order-history-2018-0.html"), "r", encoding="utf-8") as f:
             resp3 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
@@ -118,6 +118,7 @@ class TestSession(UnitTestCase):
 
         # THEN
         self.assertTrue(self.amazon_session.is_authenticated)
+        self.assertEqual(1, input_mock.call_count)
         self.assertEqual(1, resp1.call_count)
         self.assertEqual(1, resp2.call_count)
         self.assertEqual(1, resp3.call_count)
@@ -136,21 +137,21 @@ class TestSession(UnitTestCase):
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-new-otp.html"), "r", encoding="utf-8") as f:
             resp2 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-mfa.html"), "r", encoding="utf-8") as f:
             resp3 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
         with open(os.path.join(self.RESOURCES_DIR, "order-history-2018-0.html"), "r", encoding="utf-8") as f:
             resp4 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
@@ -160,6 +161,7 @@ class TestSession(UnitTestCase):
 
         # THEN
         self.assertTrue(self.amazon_session.is_authenticated)
+        self.assertEqual(2, input_mock.call_count)
         self.assertEqual(1, resp1.call_count)
         self.assertEqual(1, resp2.call_count)
         self.assertEqual(1, resp3.call_count)
@@ -177,7 +179,7 @@ class TestSession(UnitTestCase):
             )
         resp2 = responses.add(
             responses.POST,
-            "{}/ap/signin".format(BASE_URL),
+            SIGN_IN_URL,
             status=302,
             headers={"Location": "{}/ap/cvf/request".format(BASE_URL)}
         )
@@ -213,6 +215,7 @@ class TestSession(UnitTestCase):
         self.assertEqual(1, resp2.call_count)
         self.assertEqual(1, resp3.call_count)
         self.assertEqual(1, resp4.call_count)
+        # TODO: we should assert this POST uses the `data` field
         self.assertEqual(1, resp5.call_count)
 
     @responses.activate
@@ -228,7 +231,7 @@ class TestSession(UnitTestCase):
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-captcha-2.html"), "r", encoding="utf-8") as f:
             resp2 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
@@ -256,6 +259,7 @@ class TestSession(UnitTestCase):
         self.assertEqual(1, resp1.call_count)
         self.assertEqual(1, resp2.call_count)
         self.assertEqual(1, resp3.call_count)
+        # TODO: we should assert this GET uses the `params` field
         self.assertEqual(1, resp4.call_count)
 
     @unittest.skipIf(sys.platform == "win32", reason="Windows does not respect PIL's show() method in tests")
@@ -272,7 +276,7 @@ class TestSession(UnitTestCase):
             )
         resp2 = responses.add(
             responses.POST,
-            "{}/ap/signin".format(BASE_URL),
+            SIGN_IN_URL,
             status=302,
             headers={"Location": "{}/ap/cvf/request".format(BASE_URL)}
         )
@@ -304,6 +308,7 @@ class TestSession(UnitTestCase):
 
         # THEN
         self.assertTrue(self.amazon_session.is_authenticated)
+        self.assertEqual(1, input_mock.call_count)
         self.assertEqual(1, resp1.call_count)
         self.assertEqual(1, resp2.call_count)
         self.assertEqual(1, resp3.call_count)
@@ -324,7 +329,7 @@ class TestSession(UnitTestCase):
         with open(os.path.join(self.RESOURCES_DIR, "post-signin-captcha-otp.html"), "r", encoding="utf-8") as f:
             resp2 = responses.add(
                 responses.POST,
-                "{}/ap/signin".format(BASE_URL),
+                SIGN_IN_URL,
                 body=f.read(),
                 status=200,
             )
@@ -341,6 +346,7 @@ class TestSession(UnitTestCase):
 
         # THEN
         self.assertTrue(self.amazon_session.is_authenticated)
+        self.assertEqual(1, input_mock.call_count)
         self.assertEqual(1, resp1.call_count)
         self.assertEqual(1, resp2.call_count)
         self.assertEqual(1, resp3.call_count)